channel_script/requirements.txt
psycopg2-binary==2.9.3
gspread==5.3.2
xmltodict==0.13.0
